requirements.txt
pyyaml
six
requests
pillow
numpy
scipy
pandas
xarray
NetCDF
pytables
biopython
matplotlib
